VIS-181,VIS-182 separate completed/ongoing cases
