--- src/views/CasesList.vue
@@ -1,541 +1,0 @@
-<template>
-  <side-bar></side-bar>
-  <loading v-if="isLoading" :startPosition="200"></loading>
-  <div id="cases">
-    <h2>Cases</h2>
-    <div class="stats">
-      <div @click="switchCases(false)" :class="['stats-card','pointer',{'selected': completedCases === false}]">
-        <div class="column">
-          <small>Ongoing cases</small>
-          <div class="row">
-            <h4> {{table.ongoingRows.length}}</h4>        
-          </div>
-        </div>
-        <ion-icon name="albums"></ion-icon>
-      </div>
-  
-      <div @click="switchCases(true)" :class="['stats-card','pointer',{'selected': completedCases}]">
-        <div class="column">
-          <small> Completed cases</small>
-          <div class="row">
-            <h4> {{table.completedRows.length}}</h4>
-          </div>
-        </div>
-        <ion-icon name="albums"></ion-icon>
-      </div>
-
-      <div v-if="parameters.kpi" class="stats-card column">
-          <p>KPI</p>
-          <h3 class="blue">{{ parameters.kpi.value }} {{ parameters.kpi.unit }}</h3>
-          <small>Case {{ parameters.kpi.column }}  {{ parameters.kpi.operator }}</small>
-        </div>
-    </div> 
-    
-    <div v-if="completedCases" id="chart" class="shadow">
-      <h3>Recommendations history</h3>
-      <apexchart type="pie" width="380" :options="pieChart.chartOptions" :series="pieChart.series"></apexchart>
-    </div>
-    
-    <div class="cases-filter column">
-      <div @click="filterActive = !filterActive" class="filter-toggle shadow">Filters <ion-icon :class="{active : filterActive}" name="chevron-up"></ion-icon></div>
-      
-      <div class="filters column shadow" :class="{active : filterActive}">
-        <div class="row" >
-  
-          <div class="filter-component">
-            <h4 class="blue">Recommendations</h4>
-            <select v-model="filters.recommendations">
-              <option :value=true>Available</option>
-              <option :value=false>Unavailable</option>
-            </select>
-          </div>
-          <div class="filter-component">
-            <h4 class="blue">Intervened</h4>
-            <select v-model="filters.intervened">
-              <option>Yes</option>
-              <option>No</option>
-            </select>
-          </div>
-        </div>
-
-        <div class="row">
-          <button @click="filterCases" class="btn-blue">filter</button>
-          <button @click="clearFilters()" class="btn-blue">clear filters</button>
-        </div>
-      </div>
-      
-    </div>
-
-    <div v-if="filterButtonPressed" class="applied-filters row">
-      <div class="applied-filter shadow" v-for="key in appliedFilters" :key="key">
-        {{ key }} <ion-icon @click="clearFilters(key)" name="close"></ion-icon>
-      </div>
-    </div>
-
-    <div class="cases-table">
-      <table-lite
-      :is-loading="table.isLoading"
-      :page-size="table.pageSize"
-      :is-slot-mode="true"
-      :columns="table.headers"
-      :rows="table.rows"
-      :rowClasses="getRowClasses"
-      :total="table.totalRecordCount"
-      :sortable="table.sortable"
-      @do-search="doSearch"
-      @row-clicked="rowClicked"
-      @is-finished="table.isLoading = false"
-      >
-      <template v-slot:id="data">
-        <p>{{ formatId(data.value.id) }}</p>
-      </template>
-      <template v-slot:recommendations="data">
-        <div class="case-recommendations" :class="[data.value.recommendations ? 'available' : 'unavailable']"> 
-          {{ data.value.recommendations ? "recommendations available" : "no new recommendations" }}
-        </div>
-      </template>
-
-      <template v-slot:performance="data">
-        <p>{{ data.value.performance.value}} {{ data.value.performance.unit }}</p>
-      </template>
-
-      <template v-slot:outcome="data">
-        <p>{{ data.value.outcome ? 'Yes': 'No'}}</p>
-      </template>
-    
-    </table-lite>
-      
-    </div>
-    </div>
-  </template>
-  
-  <script>
-  
-  import casesService from "@/services/cases.service";
-  import logsService from "@/services/logs.service";
-  import SideBar from '@/components/SideBar.vue';
-  import Loading from "@/components/LoadingComponent.vue";
-  import TableLite from "vue3-table-lite";
-  import shared from '@/services/shared';
-  import VueApexCharts from 'vue3-apexcharts';
-
-  export default {
-    name: 'CasesList',
-  
-    components: {
-          SideBar,
-          Loading,
-          TableLite,
-          apexchart: VueApexCharts,
-        },
-
-    computed: {
-      appliedFilters(){
-        return Object.keys(this.filters).filter(key => this.filters[key] !== null);
-      }
-    },
-  
-    data() {
-      return {
-        timer: null,
-        logStatus: 'NULL',
-
-        isLoading: false,
-        cases: [],
-        casesData: [],
-
-        parameters: {},
-        
-        performanceColumn: undefined,
-        completedCases: false,
-
-        filterActive: true,
-        filterButtonPressed: false,
-
-        filters:{
-          recommendations: null,
-          intervened: null,
-          
-        },
-
-        pieChart: {
-          series: [0, 0, 1],
-          chartOptions: {
-            colors: ['#17ad37','#F5222D','#a0a3a5'],
-            chart: {
-              animations:{
-                enabled: false,
-              },
-              width: 380,
-              type: 'pie',
-            },
-            labels: ['Was successful', 'Was unsuccessful', 'No data'],
-            responsive: [{
-              breakpoint: 480,
-              options: {
-                chart: {
-                  width: 200
-                },
-                legend: {
-                  position: 'right'
-                }
-              }
-            }]
-          },
-        },
-
-        table: {
-          pageSize: 10,
-          isLoading: false,
-          headers: [
-            {
-              label: 'Case ID',
-              field: 'id',
-              width: "10%",
-              sortable: true,
-            },
-            {
-              label: "Recommendations",
-              field: "recommendations",
-              width: "5%",
-              sortable: false,
-            },
-            {
-              label: "Performance",
-              field: "performance",
-              width: "5%",
-              sortable: true,
-            },
-            {
-              label: "Intervened",
-              field: "intervened",
-              width: "5%",
-              sortable: true,
-            },
-          ],
-          outcomeHeader: {
-              label: "Positive outcome",
-              field: "outcome",
-              sortable: true,
-            },
-          rows: [],
-          ongoingRows: [],
-          completedRows:[],
-          sortable: {
-            order: shared.getLocal('casesListOrder'),
-            sort:  shared.getLocal('casesListSort')
-          },
-          totalRecordCount: 0,
-          clickedRows: [],
-        },
-      };
-    },
-  
-    mounted() {
-      if (shared.getLocal('logId')) {
-        this.table.clickedRows = shared.getLocal(`casesListClickedRows${shared.getLocal('logId')}`) || [];
-        this.isLoading = true;
-        this.getParameters();
-        this.getProjectStatus();
-      }
-    },
-
-    beforeUnmount(){
-      clearInterval(this.timer);
-    },
-    
-    methods: {
-      getRowClasses(row){
-        let index = this.table.clickedRows.indexOf(row.id);
-        if (index < 0) return '';
-        if (index === 0) return 'last-clicked-row';
-        return 'clicked-row';
-      },
-
-      formatId(id){
-        if(!id) return null;
-        return id.slice(id.indexOf('-') + 1);
-      },
-  
-      rowClicked(row){
-        let index = this.table.clickedRows.indexOf(row.id);
-        if(index > -1) {
-          this.table.clickedRows.splice(index, 1);
-        }
-        this.table.clickedRows.unshift(row.id);
-
-        shared.setLocal(`casesListClickedRows${shared.getLocal('logId')}`,this.table.clickedRows,1);
-        this.$router.push({name: 'case',params: {'caseId': row.id}})
-      },
-  
-      doSearch(offset,limit,order,sort,doSort=false,isLoading=true){
-        if (order === null || sort === null) return;
-
-        this.table.isLoading = isLoading;
-        setTimeout(() => {
-          let tempRows;
-          
-          doSort = order !== this.table.sortable.order || sort !== this.table.sortable.sort || doSort;
-          
-          this.table.sortable.order = order;
-          this.table.sortable.sort = sort;
-
-          if (this.completedCases) {
-            tempRows = doSort ? this.sortCases(this.table.completedRows) : this.table.completedRows;
-          } else {
-            tempRows = doSort ? this.sortCases(this.table.ongoingRows) : this.table.ongoingRows;
-          }
-
-          if (this.completedCases) this.table.completedRows = tempRows;
-          else this.table.ongoingRows = tempRows;
-
-          this.table.rows = tempRows.slice(offset,offset+limit);
-          this.table.totalRecordCount = tempRows.length;
-          tempRows = null;
-
-          shared.setLocal('casesListOrder',order,5);
-          shared.setLocal('casesListSort',sort,5);
-          shared.setLocal('casesListOffset',offset,5);
-          shared.setLocal('casesListLimit',limit,5);
-          
-        }, 600);
-      },
-
-      sortCases(rows){
-        let sort = this.table.sortable.sort;
-        let order = this.table.sortable.order;
-        const sortOrder = sort === 'asc' ? 1 : -1;
-        if (order === "performance" && this.performanceColumn === "DURATION"){
-          rows = rows.sort((a, b) => 
-          (shared.parseDuration(a[order]) > shared.parseDuration(b[order])) ? (1 * sortOrder) : (-1 * sortOrder));
-        } else if (order === "performance"){
-          rows = rows.sort((a, b) => 
-          (a[order].value > b[order].value) ? (1 * sortOrder) : (-1 * sortOrder));
-        }
-        else{
-          rows = rows.sort((a, b) => (a[order] > b[order]) ? (1 * sortOrder) : (-1 * sortOrder));
-        }
-        return rows;
-      },
-
-      createPieChart(){
-        let outcomeCounts = shared.groupByAndCount(this.table.completedRows,'outcome','intervened');
-        if(Object.keys(outcomeCounts).length > 0){
-          const propertiesToCheck = [true, false, null];
-          this.pieChart.series = propertiesToCheck.map(property => {
-            return outcomeCounts[property] || 0;
-          });
-        }
-      },
-
-      getParameters(){
-          logsService.getParameters(shared.getLocal('logId')).then(
-            (response) => {
-              this.parameters = response.data.parameters;
-              this.getCases();
-            },
-            (error) => {
-              this.isLoading = false;
-              const resMessage =
-                (error.response &&
-                  error.response.data &&
-                  error.response.data.error) ||
-                error.message ||
-                error.toString();
-                this.$notify({
-                        title: 'An error occured',
-                        text: resMessage,
-                        type: 'error'
-                    }) 
-            }
-          );
-        },
-      
-      getCases() {
-        casesService.getCasesByLog(shared.getLocal('logId')).then(
-          (response) => {
-            this.cases = response.data.cases;
-            if (this.cases.length > 0) this.formatCases();
-            else this.isLoading = false;
-            },
-          (error) => {
-            this.isLoading = false;
-            if (error.response.status === 504) return;
-            const resMessage =
-              (error.response &&
-                error.response.data &&
-                error.response.data.error) ||
-              error.message ||
-              error.toString();
-              this.$notify({
-                          title: 'An error occured',
-                          text: resMessage,
-                          type: 'error'
-                      }) 
-          }
-        );
-      },
-  
-      formatCases(){
-        let isLoading = true;
-        if (this.casesData.length > 0){
-          isLoading = false;
-          this.casesData = [];
-          this.table.ongoingRows = [];
-          this.table.completedRows = [];
-        } else{
-          Object.keys(this.cases[0].case_attributes).forEach(k => {
-          this.table.headers.push({
-            label:k,
-            field:k,
-            sortable:true
-          });
-        })
-      }
-      
-      let data = {};
-        for (const el of this.cases) {
-          data = this.formatCase(el);
-          this.casesData.push(data);
-        }  
-        this.table.headers[2].label = this.performanceColumn;
-        
-        this.casesData.reduce((acc, row) => {
-          if (row.completed === false) {
-            acc[0].push(row);
-          } else {
-            acc[1].push(row);
-          }
-          return acc;
-        }, [this.table.ongoingRows, this.table.completedRows]);
-
-        this.doSearch(0,this.table.pageSize,this.table.sortable.order,this.table.sortable.sort,true,isLoading);
-
-        this.createPieChart();
-
-        this.cases = null;
-        this.isLoading = false;
-      },
-  
-      formatCase(singleCase){
-          if (!this.performanceColumn && singleCase.case_performance.column !== null) this.performanceColumn = singleCase.case_performance.column;
-          
-          let caseActivities = singleCase.activities;
-          let intervened = "No";
-          let performance, outcome;
-          if(singleCase.case_performance){
-            performance = {value:singleCase.case_performance.value,unit:singleCase.case_performance.unit};
-            outcome = singleCase.case_performance.outcome
-          } else{
-            performance = {value: null,unit:null};
-            outcome = null
-          }
-  
-          caseActivities.forEach(activity => {
-            if(activity.prescriptions){
-              activity.prescriptions.forEach(prescription => {
-                if (prescription.status === 'accepted'){
-                  intervened = "Yes";
-                  return;
-                }
-              })
-            }
-          });
-          let data = {
-            id: singleCase._id, 
-            recommendations: caseActivities[caseActivities.length-1].prescriptions.length === 0 ? false : true,
-            intervened: intervened,
-            performance: performance, 
-            outcome: outcome
-          }
-          
-          Object.keys(singleCase.case_attributes).forEach(k => {
-            data[k] = singleCase.case_attributes[k];
-          })
-  
-          data['completed'] = singleCase.case_completed;
-          return data;
-      },
-  
-      switchCases(status){
-        if (this.completedCases === status){
-          return;
-        }
-        this.completedCases = status;
-        if(status) {
-          this.table.headers.splice(4,0,this.table.outcomeHeader);
-        }
-        else {
-          this.table.headers.splice(4,1);
-        }
-        this.doSearch(0,this.table.pageSize,this.table.sortable.order,this.table.sortable.sort,true);
-      },
-
-      getProjectStatus(){
-        logsService.getProjectStatus(shared.getLocal('logId')).then(
-            (response) => {
-                let status = response.data.status;
-
-                if(status === 'SIMULATING'){
-                  this.timer = setInterval(() => {
-                      this.getCases();
-                  }, 5000);
-                }
-            },
-            (error) => {
-                console.log(error);
-            }
-        ); 
-    },
-
-    doSearchAndSort(rows){
-      let offset = shared.getLocal('casesListOffset');
-      let limit = shared.getLocal('casesListLimit');
-      rows = this.sortCases(rows);
-      this.table.rows = rows.slice(offset,offset + limit);
-      this.table.totalRecordCount = rows.length;
-    },
-
-      filterCases(){
-        if (this.appliedFilters.length < 1) return;
-        this.table.isLoading = true;
-        
-        this.filterButtonPressed = true;
-        shared.setLocal("appliedFilters", this.filters);
-        
-        let tempRows = this.completedCases ? this.table.completedRows : this.table.ongoingRows;
-        
-        if(this.filters.recommendations !== null){
-          tempRows = tempRows.filter(r => r.recommendations === this.filters.recommendations);
-        }
-        if(this.filters.intervened !== null){
-          tempRows = tempRows.filter(r => r.intervened === this.filters.intervened);
-        }
-        this.doSearchAndSort(tempRows);
-        tempRows = null;
-        this.table.isLoading = false;
-      },
-      
-      clearFilters(key = false){
-        if (this.appliedFilters.length < 1) return;
-
-        this.table.isLoading = true;
-        let tempRows = this.completedCases ? this.table.completedRows : this.table.ongoingRows;
-        this.filterButtonPressed = false;
-        if (key){
-          tempRows.filter(r => r[key] !== this.filters[key])
-          this.filters[key] = null;
-          tempRows = tempRows.concat(this.table.rows);
-        } else{
-          this.filters.intervened = null;
-          this.filters.recommendations = null;
-        }
-        this.doSearchAndSort(tempRows);
-        tempRows = null;
-        this.table.isLoading = false;
-      }
-
-    }
-  
-  }
-  </script>
